pyproject.toml
[project]
name = "flowfit-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.20",
    "openai==1.78.0",
    "pg8000==1.31.2",
    "python-dotenv==1.1.0",
    "scikit-learn>=1.8.0",
    "openpyxl==3.1.5",
    "olefile>=0.47",
    "python-docx>=1.2.0",
    "pypdf>=6.10.0",
    "curl-cffi>=0.15.0",
    "yfinance>=1.2.2",
    "python-pptx>=1.0.2",
]

